Make IText support accent-based languages
keep the `keypress` event handler for normal input scenarios.

using the fact that input actions in accent-based languages will always
triggers `keydown` with keyCode `229` and following event `input`,
record the latest keyCode down and handle accent-based input separately.

In the case of accent-based input methods, there's no predictable
results according to a given keystroke, the length might be
`n` characters longer/shorter or even unchanged, the content might be
entirely re-written or changed partially, so we need to scan
`hiddenTextArea.value` and `this.text` from both sides to extract the
'changed parts' and the range, then apply it on IText.

one major change except the input handling part is:
`hiddenTextArea.value`, `hiddenTextArea.selectionStart` and
`hiddenTextArea.selectionEnd` should not be touched during these input actions,
otherwise the input process & state of IME will be intercepted. So an
additional parameter `noTextAreaUpdate` is added to several methods to
determine if the opperation is triggered by accent-based languages
input.